refactor: 프로그램 기능 선택 controller로 이전
