+++ src/main/java/pairmatching/domain/matching/Matching.java
@@ -1,5 +1,7 @@
 package pairmatching.domain.matching;
 
+import java.util.List;
+
 import pairmatching.domain.input.Course;
 import pairmatching.domain.input.Level;
 import pairmatching.domain.mission.Mission;
@@ -8,4 +10,5 @@ public class Matching {
 	private Course course;
 	private Level level;
 	private Mission mission;
+	private List<Pair> matchedPairs;
 }
