+++ src/main/java/pairmatching/domain/matching/Pair.java
@@ -1,0 +1,11 @@
+package pairmatching.domain.matching;
+
+import java.util.List;
+
+public class Pair {
+	private List<String> pair;
+
+	public Pair(List<String> pair) {
+		this.pair = pair;
+	}
+}
